move outstation helper code to 'details' sub-module
