+++ dnp3/src/outstation/details/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod control;
+pub(crate) mod prefix;
